fix: move twind config file from `landing-page/twind/config.ts` to `landing-page/twind.config.ts`
this helps the twind-intellisense plugin a bit as it can locate the
config file
