+++ landing-page/pages/_document.tsx
@@ -1,6 +1,6 @@
 import Document, { Html, Head, Main, NextScript } from 'next/document'
 import withNextDocument from '../twind/next/document'
-import { config } from '../twind/config'
+import { config } from '../twind.config'
 
 class MyDocument extends Document {
   render() {
